flux_requirements.txt
numpy==1.26.3
pillow==10.2.0
torch==2.0.1
transformers==4.45.2
diffusers==0.30.1
torchvision==0.15.2
einops==0.6.1
opencv-python==4.7.0.68
xformers==0.0.21
gradio==3.36.1
insightface==0.7.3
accelerate==1.0.0
peft==0.8.2
onnxruntime-gpu==1.16.0
timm==0.6.12
controlnet-aux==0.0.7
sentencepiece==0.1.99
openai-clip==1.0.1
protobuf==3.20.3